added an error message to throw in acse of drop_Schema_Cascade

+++ plugin/trino-teradata/src/main/java/io/trino/plugin/teradata/TeradataClient.java
@@ -397,6 +397,9 @@ public class TeradataClient
     {
 //        String deleteSchema = "DELETE DATABASE " + quoted(remoteSchemaName);
 //        execute(session, connection, deleteSchema);
+        if (cascade) {
+            throw new TrinoException(NOT_SUPPORTED, "This connector does not support dropping schemas with CASCADE option");
+        }
         String dropSchema = "DROP DATABASE " + quoted(remoteSchemaName);
         execute(session, connection, dropSchema);
     }
